use int64 in status messages

+++ internal/pkg/api/models/v1/remote_cluster.go
@@ -13,7 +13,7 @@ const (
 // StatusDistribution is a status and count pair used by the remote cluster status endpoint.
 type StatusDistribution struct {
 	Status string `json:"status"`
-	Count  int    `json:"count"`
+	Count  int64  `json:"count"`
 }
 
 // RemoteCluster is a standalone or clustered LXD cluster.
@@ -21,20 +21,20 @@ type RemoteCluster struct {
 	Name               string               `json:"name"`
 	Description        string               `json:"description"`
 	ClusterCertificate string               `json:"cluster_certificate"`
-	DiskThreshold      int                  `json:"disk_threshold"`
-	MemoryThreshold    int                  `json:"memory_threshold"`
+	DiskThreshold      int64                `json:"disk_threshold"`
+	MemoryThreshold    int64                `json:"memory_threshold"`
 	Status             string               `json:"status"`
-	CPUTotalCount      int                  `json:"cpu_total_count"`
+	CPUTotalCount      int64                `json:"cpu_total_count"`
 	CPULoad1           string               `json:"cpu_load_1"`
 	CPULoad5           string               `json:"cpu_load_5"`
 	CPULoad15          string               `json:"cpu_load_15"`
-	MemoryTotalAmount  int                  `json:"memory_total_amount"`
-	MemoryUsage        int                  `json:"memory_usage"`
-	DiskTotalSize      int                  `json:"disk_total_size"`
-	DiskUsage          int                  `json:"disk_usage"`
-	MemberCount        int                  `json:"member_count"`
+	MemoryTotalAmount  int64                `json:"memory_total_amount"`
+	MemoryUsage        int64                `json:"memory_usage"`
+	DiskTotalSize      int64                `json:"disk_total_size"`
+	DiskUsage          int64                `json:"disk_usage"`
+	MemberCount        int64                `json:"member_count"`
 	MemberStatuses     []StatusDistribution `json:"member_statuses"`
-	InstanceCount      int                  `json:"instance_count"`
+	InstanceCount      int64                `json:"instance_count"`
 	InstanceStatuses   []StatusDistribution `json:"instance_statuses"`
 	UIURL              string               `json:"ui_url"`
 	JoinedAt           time.Time            `json:"joined_at"`
@@ -46,8 +46,8 @@ type RemoteCluster struct {
 type RemoteClusterPatch struct {
 	Status          RemoteClusterStatus `json:"status"`
 	Description     string              `json:"description,omitempty"`
-	DiskThreshold   int                 `json:"disk_threshold,omitempty"`
-	MemoryThreshold int                 `json:"memory_threshold,omitempty"`
+	DiskThreshold   int64               `json:"disk_threshold,omitempty"`
+	MemoryThreshold int64               `json:"memory_threshold,omitempty"`
 }
 
 // RemoteClusterPost represents the fields required to create a new cluster.
@@ -59,14 +59,14 @@ type RemoteClusterPost struct {
 
 // RemoteClusterStatusPost is sent by LXD with to inform about its current status.
 type RemoteClusterStatusPost struct {
-	CPUTotalCount     int                  `json:"cpu_total_count"`
+	CPUTotalCount     int64                `json:"cpu_total_count"`
 	CPULoad1          string               `json:"cpu_load_1"`
 	CPULoad5          string               `json:"cpu_load_5"`
 	CPULoad15         string               `json:"cpu_load_15"`
-	MemoryTotalAmount int                  `json:"memory_total_amount"`
-	MemoryUsage       int                  `json:"memory_usage"`
-	DiskTotalSize     int                  `json:"disk_total_size"`
-	DiskUsage         int                  `json:"disk_usage"`
+	MemoryTotalAmount int64                `json:"memory_total_amount"`
+	MemoryUsage       int64                `json:"memory_usage"`
+	DiskTotalSize     int64                `json:"disk_total_size"`
+	DiskUsage         int64                `json:"disk_usage"`
 	MemberStatuses    []StatusDistribution `json:"member_statuses"`
 	InstanceStatuses  []StatusDistribution `json:"instance_statuses"`
 	Metrics           string               `json:"metrics"`
